+++ src/test/java/model/state/finished/StayTest.java
@@ -45,25 +45,25 @@ class StayTest {
     @ParameterizedTest
     @DisplayName("딜러의 카드가 블랙잭이거나 자신의 카드 합이 딜러의 카드 합보다 작으면 -1.0배, 자신의 카드합이 딜러의 카드합보다 높으면 1.0배, 합이 같다면 0을 반환한다.")
     @MethodSource("provideDealerCardsAndExpectedProfit")
-    void profit(final Cards dealerCards, final double expectedProfit) {
+    void profit(final Cards dealerCards, final int expectedProfit) {
         int bettingMoney = 10000;
-        double actualProfit = state.profit(bettingMoney, dealerCards);
+        int actualProfit = state.profit(bettingMoney, dealerCards);
         assertThat(actualProfit).isEqualTo(expectedProfit);
     }
 
     private static Stream<Arguments> provideDealerCardsAndExpectedProfit() {
         return Stream.of(
                 Arguments.of(Cards.from(Arrays.asList(
-                        Card.of(10, 2), Card.of(1, 2))), -10000.0),
+                        Card.of(10, 2), Card.of(1, 2))), -10000),
                 Arguments.of(Cards.from(Arrays.asList(
-                        Card.of(10, 2), Card.of(8, 3))), 0.0),
+                        Card.of(10, 2), Card.of(8, 3))), 0),
                 Arguments.of(Cards.from(Arrays.asList(
-                        Card.of(10, 3), Card.of(9, 2))), -10000.0),
+                        Card.of(10, 3), Card.of(9, 2))), -10000),
                 Arguments.of(Cards.from(Arrays.asList(
-                        Card.of(10, 3), Card.of(7, 2))), 10000.0),
+                        Card.of(10, 3), Card.of(7, 2))), 10000),
                 Arguments.of(Cards.from(Arrays.asList(
                         Card.of(10, 3),
-                        Card.of(5, 3), Card.of(7, 4))), 10000.0)
+                        Card.of(5, 3), Card.of(7, 4))), 10000)
         );
     }
 }
